+++ packages/core/src/observable/index.ts
@@ -1,5 +1,6 @@
-export { isObservable, observable } from './observable';
 export type { inferObservableValue } from './observable';
+export { isObservable, observable, observableToPromise } from './observable';
+export { map, share, tap } from './operators';
 export type {
   Observable,
   Observer,
@@ -7,5 +8,3 @@ export type {
   Unsubscribable,
   UnsubscribeFn,
 } from './types';
-export { map, share, tap } from './operators';
-export { observableToPromise } from './internals/observableToPromise';
